adding role admin and user, making management event is authorized to admin only

+++ app/GraphQL/Mutations/EventMutation.php
@@ -4,9 +4,20 @@ namespace App\GraphQL\Mutations;
 
 use App\Models\Event;
 use illuminate\Support\Facades\auth;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Auth\Access\AuthorizationException;
 
 class EventMutation{
     public function createEvent($root, array $args){
+
+        // Check if the user is authenticated
+        if (!auth()->check()) {
+            throw new AuthorizationException('You must be logged in to create an event.');
+        }
+        // Check if the user is an admin
+        if (Auth::user()->role !== 'admin') {
+        throw new \Exception('Unauthorized: Only admins can create events.');
+        }
         $event = Event::create([
             'title' => $args['title'],
             'description' => $args['description'],
@@ -28,6 +39,14 @@ class EventMutation{
         return $event;
     }
     public function deleteEvent($root, array $args){
+        // Check if the user is authenticated
+        if (!auth()->check()) {
+            throw new AuthorizationException('You must be logged in to delete an event.');
+        }
+        // Check if the user is an admin
+        if (Auth::user()->role !== 'admin') {
+            throw new \Exception('Unauthorized: Only admins can delete events.');
+        }
         $event = Event::findOrFail($args['id']);
         $event->delete();
